P10_01_chatbot/requirements.txt
botbuilder-core==4.16.2
botbuilder-dialogs==4.16.2
botbuilder-ai==4.16.2
botbuilder-applicationinsights==4.16.2
botbuilder-integration-aiohttp==4.16.2
aiohttp==3.9.5
opencensus-ext-azure==1.1.13
azure-ai-language-conversations==1.1.0
azure-cognitiveservices-language-luis==0.7.0
python-dotenv==1.0.1
pandas==2.2.2
pytest==8.2.2
pytest-asyncio==0.23.7
aiounittest==1.4.2
